Fix the reading stderr and stdout
The error condition in the while loop was impossible to hit. If we
ignore the EINTR, which we actually are allowed to do in non-blocking
mode if I read the specs correctly, the error handling could actually be
placed outside the loop.

I also corrected the address of the buffer. An array of a type already is
a pointer to the first element. Technically it worked the way it was,
but it's not the standard way of writing this.