Pass `cmd_stdout` and `cmd_stderr` as binary data
On the DB side, `cmd_stdout` and `cmd_stderr` were always already of
type `bytea`.  This change allowed me to get rid of a `convert_to()`
hack in the prepared `UPDATE` statement, so that `UPDATE`ing no longer
only works with UTF-8-encoded text data, but now also with … any
`bytea` data, as it was always intended to work.

What's still missing is to retrieve the `cmd_stdin` as binary data.
This is due to a limitation in `libpq`; from the
[docs](https://www.postgresql.org/docs/current/libpq-exec.html#LIBPQ-PQEXECPARAMS):

> There is not currently a provision to obtain different result columns
> in different formats, although that is possible in the underlying
> protocol.