Print stderr progress and again on error
We don't have 'stdlog', which makes it a bit confusing, but now the idea
is:

A command's stderr is printed as NOTICE as it's running. On unsuccessful
exit, it's repeated as ERROR, marked with a BEGIN and END line.

Lines logged line by line, to properly prefix them with the name of the
queue they came out of.