Fix quiting by signal
It happened to me several times that the daemon wouldn't quit.

There were several issues:

1) The pipe through which we throw the signal number can't be
non-blocking, because we have no opportunity to go back to the signal
handler if we would abort prematurely.

2) The syscall interrupt errno was erroneously compared to SIGINT.

3) The result of writing to the pipe was stored in an unsigned size_t,
meaning the while loop would not abort on error. This was observed in
the debugger, where kill_pipe_bytes_written was overflown to a very high
possitive number.

4) The read of the signal number from the pipe had a stall condition,
which was only not a problem because the signal handler's write loop
wrote too often.